moved store state types into store modules
