--- packages/renderer/src/store/types/ApplicationState.ts
@@ -1,10 +1,0 @@
-import type { Column, DropZone } from "interfaces";
-
-type ApplicationState = {
-  isSidebarMinimized: boolean;
-  activeView: string;
-  dropZones: Array<DropZone>;
-  columns: Array<Column>;
-};
-
-export default ApplicationState;
