--- packages/renderer/src/store/types/ProjectState.ts
@@ -1,8 +1,0 @@
-import type { Project } from "interfaces";
-
-type ProjectState = {
-  all: Array<Project>;
-  active: Project | null;
-};
-
-export default ProjectState;
